segregated photo services from album services
- added photoRepository and photoServices class to handle photo related services
