+++ Gallery/Gallery/Domain/Repository/PhotosRepository.swift
@@ -1,0 +1,7 @@
+class PhotosRepository {
+    let photosServices = PhotosServices()
+    func getPhotosForAlbum(albumId: Int, completion : @escaping([PhotoEntityModel]?,Error?) -> ()) {
+        let photosUrl = Constants.photosUrl+"\(albumId)/photos"
+        photosServices.getPhotosForAlbum(urlString: photosUrl, completion: completion)
+    }
+}
